unified logging behaviour
